+++ src/PimEnterprise/Bundle/EnrichBundle/Form/View/ProductFormView.php
@@ -11,10 +11,11 @@
 
 namespace PimEnterprise\Bundle\EnrichBundle\Form\View;
 
+use PimEnterprise\Bundle\SecurityBundle\Attributes;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\EnrichBundle\Form\View\ProductFormView as BaseProductFormView;
-use PimEnterprise\Bundle\SecurityBundle\Attributes;
+use Pim\Bundle\EnrichBundle\Form\View\ViewUpdater\ViewUpdaterRegistry;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\Security\Core\SecurityContextInterface;
 
@@ -33,10 +34,13 @@ class ProductFormView extends BaseProductFormView
     /**
      * Construct
      *
+     * @param ViewUpdaterRegistry      $viewUpdaterRegistry
      * @param SecurityContextInterface $securityContext
      */
-    public function __construct(SecurityContextInterface $securityContext)
+    public function __construct(ViewUpdaterRegistry $viewUpdaterRegistry, SecurityContextInterface $securityContext)
     {
+        parent::__construct($viewUpdaterRegistry);
+
         $this->securityContext = $securityContext;
     }